logging-service: split frontend-collector and logging library
