+++ docker-compose/.env.example
@@ -117,7 +117,7 @@ EXCEL_LOG_LEVEL=info
 EXPORT_HOST=excel-export-service
 EXPORT_PORT=8888
 
-# logging-service
+# frontend-collector
 LOGGER_PORT=3001
 LOG_LEVEL=info
 LOGGING_SERVICE_CACHE_DURATION=20
