+++ docs/developer/logging.md
@@ -42,6 +42,6 @@ There are two settings of the Pino logger that are set via environment variables
 
 ## Frontend Logging
 
-TruBudget offers the possibility to log frontend events such as errors or crashes. To collect and store this data, the TruBudget logging-service must be enabled.
-To enable the TruBudget logging-service, the parameter `--with-frontend-logging` must be passed when deploying or when starting TruBudget in development mode with the provided start script.
+TruBudget offers the possibility to log frontend events such as errors or crashes. To collect and store this data, the TruBudget frontend-collector must be enabled.
+To enable the TruBudget frontend-collector, the parameter `--with-frontend-logging` must be passed when deploying or when starting TruBudget in development mode with the provided start script.
 Make sure all env variables are set correctly as defined [in the environment description](./../environment-variables.md).
