+++ frontend-collector/.dockerignore
