+++ frontend-collector/.env.example
@@ -1,0 +1,4 @@
+LOGGER_PORT=3001
+LOG_LEVEL=trace
+LOGGING_SERVICE_CACHE_DURATION=20
+NODE_ENV=development
