+++ frontend-collector/.nsprc
@@ -1,0 +1,1 @@
+{}
