+++ frontend-collector/Dockerfile
@@ -1,7 +1,6 @@
-FROM node:20.9.0-alpine
+FROM node:18.4.0-alpine
 
 WORKDIR /home/node
-COPY index.js index.js
 COPY serve.js serve.js
 COPY package.json ./
 COPY package-lock.json ./
