+++ frontend-collector/README.md
@@ -1,0 +1,4 @@
+# Frontend Logging
+
+TruBudget offers the possibility to log frontend events such as errors or crashes. To collect and store this data, the TruBudget frontend-collector must be enabled.
+To enable the TruBudget frontend-collector, the parameter `--with-frontend-logging` must be passed when deploying or when starting TruBudget in development mode with the provided start script.
