+++ logging-service/.env.example
@@ -1,6 +1,2 @@
-LOGGER_PORT=3001
-API_HOST=localhost
-API_PORT=8080
+PRETTY_PRINT=true
 LOG_LEVEL=trace
-LOGGING_SERVICE_CACHE_DURATION=20
-NODE_ENV=development
