+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "danger": "danger",
     "audit": "better-npm-audit audit",
-    "update-logging": "(cd provisioning && npm i trubudget-logging-service@latest --save); (cd blockchain && npm i trubudget-logging-service@latest --save); (cd api && npm i trubudget-logging-service@latest --save); (cd email-notification-service && npm i trubudget-logging-service@latest --save); (cd excel-export-service && npm i trubudget-logging-service@latest --save); (cd storage-service && npm i trubudget-logging-service@latest --save)"
+    "update-logging": "(cd provisioning && npm i trubudget-logging-service@latest --save); (cd blockchain && npm i trubudget-logging-service@latest --save); (cd api && npm i trubudget-logging-service@latest --save); (cd email-notification-service && npm i trubudget-logging-service@latest --save); (cd excel-export-service && npm i trubudget-logging-service@latest --save); (cd storage-service && npm i trubudget-logging-service@latest --save); (cd frontend-collector && npm i trubudget-logging-service@latest --save)"
   },
   "dependencies": {
     "better-npm-audit": "^3.7.3"
